Change Entry/Revision models drastically
This makes some drastic changes to the Entry/Revision models.

1. We not longer have a ForeignKey from the Entry to the Revision using
both the Revision id and Revision number. Instead this is now broken out
using a secondary table that does a one-to-many relationship. This takes
care of a lot of things, such as not having a safe way to update what
the revision number was.
2. Entry directly links to the latest Revision for speed/ease of use.
3. The text field "entry" is deferred while loading from the DB, this
contains the raw unformatted text, in the future the revision will have
a direct link to a caching table that holds a cache of the current
revision. This will also allow easily adding beaker or dogpile in the
future to use memcache for instance.
4. Entries can now have multiple authors, revisions can only have one
author.

The idea here is that if a secondary user makes a bunch of changes the
owner (not yet sure how I am going to denote that) can approve the
changes and the user that made that changes is added as one of the
people who contributed to the entry.

Question is, do we want to allow this, or should there be some other
mechanism for inviting new people onto an existing entry. Should it be
possible for example to co-write an article and give the two authors
equal credit, but also allow editors to have their name on it? Need to
do some research to see what big publishers do.