Fixes up the yearly template
The yearly template used to expect the entries to come in as a list of
months, that itself contained a list of posts for that month. Now it
just receives a list of entries, and has to make due with them.

This changes the way the template is rendered to render the template as
required, it required some reordering of the template so that it would
correctly build the required div's/unordered lists.