Move from using a DateTime object to a composite
This attempts to fix an issue with larger datasets on Entry. pubdate was
being used heavily for storing the date that a post was being published,
yet at the same time we were also using it to filter.

Using extract(what, from) it is possible for example to extract('year',
Entry.pubdate), and then use it in a SQLAlchemy filter, the downside
is that PostgreSQL at that point is forced to do a sequential table
scan, which is log(N). That is NOT a good idea, even if the column
itself is indexed there is no way to speed it up since it has to extract
specific data from the stored value.

This composite method takes the time to do most of the work in Python,
and gives us seperate columns for year, month, day, which we can then do
full queries on, and since an index is available they are fast.

Using a composite in Python we can also get it to look to the rest of
the app like any other datetime.date object, or at least close enough to
not make a real difference.