Add some properties to Entry for easy access
Add some new properties to Entry for ease of use in templating or
end-designs. There is no good reason why anyone needs to care that the
title is stored in the current_revision and not in the Entry (this
allows for example the title change to be versioned as well).

Also checks to make sure that we are getting a datetime.datetime for
pubtime. In the future if you set pubtime you will also automatically
set pubdate and vice-versa.