Rename custom Date() class to PublishedDateTime()
This composite property takes care of using four seperate columns to
operate on rather than using a DateTime directly. The reason for doing
so is query speed. When querying data using extract() it gets expensive
fast as a full table scan has to be completed.

We also define a custom comparator that allows us to define custom SQL
to compare objects in SQL, this for instance allows us to check if a
PublishedDate is NULL or not.

We also define a custom ascending/descending sql that correctly sorts
ascending or descending based upon the year, month, day, and time
columns that we are using underneath.