Create new view/templating helper URLHelper
URLHelper is a simple class that helps build the correct URL's according
to the routes as set up. It does this using route_url, but allows for
special processing of input parameters.

This allows one to easily pass in an Entry model for instance, and
receive back a URL to that specific Entry's full URL/permapath for
example, or retrieve the URL for the year/month archives.

The idea here is to help remove some of the fragility in terms of route
naming and move it into a central location so that in the future if the
route changes hopefully no changes have to be made to the templates, and
it also solves the problem of having very long route_url calls in the
template code whereby the information is pulled from an already existing
Entry or cobbled together from variables. If an Entry exists, you can
freely use it, if not, you can give the parameters directly in the order
required (ymd_archive for example expects year, month, day as positional
arguments).