New test infrastructure for testing against pgsql
Instead of using the previous test structure of creating a new
database with the models repopulating it for each test, we want to make
use of PostgreSQL so that we can test against a real database.

This adds the required setup stuff that will create the database, start
the transactions and let the tests run, rolling back after every test is
complete.