Two routes, both with predicates...
When we go process a request, we want to see if the management URL is
part of the request, and is so we set up a different route pattern, with
eventually a different root.

This will allow us to run the same Pyramid app, yet have many URL's
pointing at it without making the management interface available on all
of them.