Fix a small bug
When passing in very large keys to the HMAC it will want to truncate it
to the size of the block for the selected hasing algorithm. This is done
by creating a hashing object and getting the return from it and using
that as the key.

This won't work if the digestmod passed in can't accept this paramater
and pass it on.