Change in forbidden_view_config in Pyramid 1.5a3
If forbidden_view_config is passed a context, it raises an error, it
used to instead override whatever the user provided with HTTPForbidden.

This behaviour has changed, and means it is no longer compatible with
the way I was using it along with @view_defaults().

Updating to use @view_config() which does work, and causes no further
side-effects.