Add APITicket object
This is used to create new API Tickets, and or fetch existing ones. It
returns the API ticket base64 encoded signed/serialized.

This fetches the serializer from the registry using queryUtility. This
is used so that in the future if we want to swap out the serializer we
can do so.